+++ spotinst/elastigroup_aws_launch_configuration/consts.go
@@ -5,7 +5,6 @@ import "github.com/spotinst/terraform-provider-spotinst/spotinst/commons"
 const (
 	Prefix = "launch_configuration_"
 )
-
 const (
 	ImageId            commons.FieldName = "image_id"
 	IamInstanceProfile commons.FieldName = "iam_instance_profile"
@@ -18,9 +17,12 @@ const (
 	PlacementTenancy   commons.FieldName = "placement_tenancy"
 	CPUCredits         commons.FieldName = "cpu_credits"
 	MetadataOptions    commons.FieldName = "metadata_options"
-
+	CPUOptions         commons.FieldName = "cpu_options"
 	// - MetadataOptions -----------------------------
 	HTTPTokens              commons.FieldName = "http_tokens"
 	HTTPPutResponseHopLimit commons.FieldName = "http_put_response_hop_limit"
+	// -----------------------------------
+	// - CPUOptions -----------------------------
+	ThreadsPerCore commons.FieldName = "threads_per_core"
 	// -----------------------------------
 )